Add files via upload
In the player-versus-player Tic-Tac-Toe game, the row and column numbers are input as integers. The software will prompt the user for input again if they attempt to enter a string, reenter the location where the symbol (X/O) is located, or enter an integer that is outside of the index boundary.