Add files via upload
The project operates as:
Numerical values are entered for the row and column numbers. If the user tries to enter a string, or goes back to where the symbol (X/O) is already present, or enters an integer that is outside the index boundary, the software will ask them for input once more.
The game concludes when one of the players successfully creates their sign in either a horizontal or vertical direction.