Add files via upload
This program contains classes for creation of - cards, deck of cards, for shuffling of deck, distribution of cards among players and checking the condition for winner declaration using concept of Python OOP using classes and objects creation . The game is played as : A deck is divided among players, with each having a down stack. In a "battle," the player with the higher card moves both cards to their stack. Aces are high, and suits are ignored. If the cards are equal, a "war" occurs. Both players face down and face up, with the higher card winning the war. If the face-up cards are equal again, the battle repeats. This repeats until one player's face-up card is higher than their opponent's.