Add files via upload
The Project  is designed for management of parking system , which takes input as vehicle name, owner name and contact number of owner. It works by taking input as, if the vehicle is two wheeler or four wheeler and charge accordingly to the owner of parking. The system does not take entry for the same vehicle by considering the name of owner and vehicle , limit is applied for number of vehicles parked and can be changed accordingly as per the need of user. Project uses 'streamlit' library for better UI along with 'lottie' and function.